% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committee_contributions.R
\docType{data}
\name{committee_contributions}
\alias{committee_contributions}
\title{Committee contributions metadata}
\format{A data frame with 1210 rows and 3 columns with information about candidate committee contibutions.
\describe{
\item{cand_id}{Recipient Candidate ID. Candidate receiving money from the filing committee}
\item{number_of_contributions}{Total number of contibutions from committees}
\item{total_contributions}{Sum of all contributions made to each candidate, in dollars}
}}
\source{
\url{https://classic.fec.gov/finance/disclosure/metadata/DataDictionaryContributionstoCandidates.shtml}, downloaded 2019-12-09
}
\usage{
committee_contributions
}
\description{
Provides data pertaining to each candidate and their contibutions from committees during the 2016 general election.
}
\keyword{datasets}
