% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individuals.R
\docType{data}
\name{individuals}
\alias{individuals}
\title{Individual Contributions Master metadata}
\format{A data frame with 5000 rows and 5 columns, providing information for each transaction:
\describe{
\item{committee_id}{A unique code assigned to a committee by the FEC}
\item{state}{Individual's state}
\item{zipcode}{Individuals's zipcode}
\item{transaction_date}{Date of transaction}
\item{transaction_amount}{Amount of transaction, in dollars}
}}
\source{
\url{https://www.fec.gov/campaign-finance-data/contributions-individuals-file-description/}, downloaded 2019-12-09
}
\usage{
individuals
}
\description{
Basic information for a sample of transactions from individuals to candidates/committees.
}
\keyword{datasets}
